refactor: extract input.c from somewm.c
Move all input handling into a dedicated module: pointer events
(axisnotify, buttonpress, motionnotify, motionrelative, motionabsolute),
keyboard (keybinding, keypress, keypressmod, keyrepeat), gestures
(swipe/pinch/hold), pointer constraints, input device management,
hit testing (xytonode, drawin_accepts_input_at), cursor helpers, and
seat request handlers (selection, drag).

45 functions, 23 listener structs, and the TrackedPointer typedef
extracted. This is the largest single extraction.

somewm.c: 5299 -> 3471 lines (-1828)
input.c: 1899 lines (new)
